Moved NoisyOr implementation to NoisyOrModel
