+++ pgmpy/estimators/NoisyOrMLE.py
@@ -11,6 +11,7 @@ from pgmpy.estimators import ParameterEstimator
 from pgmpy.estimators import MaximumLikelihoodEstimator
 from pgmpy.factors.discrete.NoisyOrCPD import NoisyOrCPD
 from pgmpy.models import BayesianModel
+from pgmpy.models import NoisyOrModel
 
 
 class NoisyOrMLE(ParameterEstimator):
@@ -155,195 +156,6 @@ class NoisyOrMLE(ParameterEstimator):
         return cpd
 
 
-def LeakyLL(theta, X, y, w):
-    """
-    Log likelihood of parent data X and child data y with leak node.
-    """
-    D = np.append(y, X, axis=1)
-    likelihoods = np.apply_along_axis(LikelihoodLeaky, 1, D, theta=theta)
-    log_likelihoods = np.log(likelihoods)
-    weighted_log_likelihoods = np.multiply(w, log_likelihoods)
-    return -np.sum(weighted_log_likelihoods)
-
-
-def NoLeakyLL(theta, X, y, w):
-    """
-    Log likelihood of parent data X and child data y without
-    leak node.
-    """
-    D = np.append(y, X, axis=1)
-    likelihoods = np.apply_along_axis(LikelihoodNoLeaky, 1, D, theta=theta)
-    log_likelihoods = np.log(likelihoods)
-    weighted_log_likelihoods = np.multiply(w, log_likelihoods)
-    return -np.sum(weighted_log_likelihoods)
-
-
-def LikelihoodLeaky(D, theta):
-    """
-    Likelihood of single data observation D with leak node.
-    y is assumed to be first element of vector D.
-    """
-    theta0 = theta[0]
-    thetaI = theta[1:]
-    y = D[0]
-    X = D[1:]
-    p = theta0 * np.prod(np.power(thetaI, X))
-    if y == 0:
-        return p
-    if y == 1:
-        return 1 - p
-    else:
-        return ValueError
-
-
-def LikelihoodNoLeaky(D, theta):
-    """
-    Likelihood of single data observation D without leak node.
-    y is assumed to be first element of vector D.
-    """
-    y = D[0]
-    X = D[1:]
-    p = np.prod(np.power(theta, X))
-    if y == 0:
-        return p
-    if y == 1:
-        return 1 - p
-    else:
-        return ValueError
-
-
-def GradientLeaky(theta, X, y, w):
-    """
-    Leaky gradient of parent data X and child data y.
-    """
-    grad = np.zeros(theta.shape)
-    grad[0] = LeakGradient(theta, X, y, w)
-    grad[1:] = InhibitorGradientLeaky(theta, X, y, w)
-    return -grad
-
-
-def GradientNoLeaky(theta, X, y, w):
-    """
-    No-leaky gradient of parent data X and child data y.
-    """
-    grad = np.zeros(theta.shape)
-    grad[:] = InhibitorGradientNoLeaky(theta, X, y, w)
-    return -grad
-
-
-def LeakGradient(theta, X, y, w):
-    """
-    Leak node gradient.
-    """
-    theta0 = theta[0]
-    thetaI = theta[1:]
-    y0_idx = np.where((y == 0).all(axis=1))
-    y1_idx = np.where((y == 1).all(axis=1))
-    n_y0 = np.sum(w[y0_idx])
-    term1 = n_y0 / theta0
-    X_y1 = X[y1_idx]
-    w_y1 = w[y1_idx]
-    try:
-        term2 = np.apply_along_axis(LeakQuotient, 1, X_y1,
-                                    theta0=theta0, thetaI=thetaI)
-        w_y1.shape = (w_y1.shape[0],)
-        term2 = sum(np.multiply(w_y1, term2))
-        return term1 - term2
-    except ValueError:
-        print('0 instances of y == 1 in data')
-        print(x_y1)
-
-
-def LeakQuotient(x, theta0, thetaI):
-    """
-    Quotient term in leak node gradient.
-    """
-    numerator = np.prod(np.power(thetaI, x))
-    denominator = 1 - theta0 * np.prod(np.power(thetaI, x))
-    return numerator / denominator
-
-
-def InhibitorGradientLeaky(theta, X, y, w):
-    """
-    Leaky gradient of inhibitor parameters,
-    excluding leak node parameter.
-    """
-    theta0 = theta[0]
-    thetaI = theta[1:]
-    grad = []
-    for i in range(X.shape[1]):
-        thetai = thetaI[i]
-        x = X[:, i]
-        x = np.reshape(x, (x.shape[0], 1))
-        D = np.append(x, y, axis=1)
-        # Subset data by x=1,y=0 & x=1,y=1
-        x1_y0_idx = np.where(np.all(D == (1, 0), axis=1))
-        x1_y1_idx = np.where(np.all(D == (1, 1), axis=1))
-        n_x1_y0 = np.sum(w[x1_y0_idx])
-        x1_y1 = X[x1_y1_idx]
-        w_x1_y1 = w[x1_y1_idx]
-        term1 = n_x1_y0 / thetai
-        term2 = np.apply_along_axis(InhibitorQuotientLeaky,
-                                    1,
-                                    x1_y1,
-                                    theta0=theta0,
-                                    thetaI=thetaI,
-                                    i=i)
-        w_x1_y1.shape = (w_x1_y1.shape[0],)
-        term2 = sum(np.multiply(w_x1_y1, term2))
-        grad.append(term1 - term2)
-    return grad
-
-
-def InhibitorGradientNoLeaky(theta, X, y, w):
-    """
-    No-leaky gradient of inhibitor parameters.
-    """
-    grad = []
-    for i in range(X.shape[1]):
-        thetai = theta[i]
-        x = X[:, i]
-        x = np.reshape(x, (x.shape[0], 1))
-        D = np.append(x, y, axis=1)
-        # Subset data by x=1,y=0 & x=1,y=1
-        x1_y0_idx = np.where(np.all(D == (1, 0), axis=1))
-        x1_y1_idx = np.where(np.all(D == (1, 1), axis=1))
-        n_x1_y0 = np.sum(w[x1_y0_idx])
-        x1_y1 = X[x1_y1_idx]
-        w_x1_y1 = w[x1_y1_idx]
-        term1 = n_x1_y0 / thetai
-        term2 = np.apply_along_axis(InhibitorQuotientNoLeaky,
-                                    1,
-                                    x1_y1,
-                                    theta=theta,
-                                    i=i)
-        term2 = sum(np.multiply(w_x1_y1, term2))
-        grad.append(term1 - term2)
-    return grad
-
-
-def InhibitorQuotientLeaky(x, theta0, thetaI, i):
-    """
-    Leaky quotient term in inhibitor gradient.
-    """
-    x_exclude = np.delete(x, i)
-    thetaI_exclude = np.delete(thetaI, i)
-    numerator = theta0 * np.prod(np.power(thetaI_exclude, x_exclude))
-    denominator = 1 - (theta0 * np.prod(np.power(thetaI, x)))
-    return numerator / denominator
-
-
-def InhibitorQuotientNoLeaky(x, theta, i):
-    """
-    No-leaky quotient term in inhibitor gradient.
-    """
-    x_exclude = np.delete(x, i)
-    theta_exclude = np.delete(theta, i)
-    numerator = np.prod(np.power(theta_exclude, x_exclude))
-    denominator = 1 - np.prod(np.power(theta, x))
-    return numerator / denominator
-
-
 def NoisyOrFit(X, y, w, leaky):
     """
     Learn NoisyOr parameters from MLE.
@@ -366,11 +178,11 @@ def NoisyOrFit(X, y, w, leaky):
 
     if leaky:
         m += 1
-        objective = LeakyLL
-        gradient = GradientLeaky
+        objective = NosiyOrModel.LeakyLL
+        gradient = NoisyOrModel.GradientLeaky
     else:
-        objective = NoLeakyLL
-        gradient = GradientNoLeaky
+        objective = NoisyOrModel.NoLeakyLL
+        gradient = NosiyOrModel.GradientNoLeaky
 
     thetaInit = 0.5 * np.ones((m,))
     bounds = Bounds(lb=0.001, ub=0.999)
